create enums
