+++ react/modules/constants.ts
@@ -1,0 +1,9 @@
+export enum Orientations {
+  horizontal = 'horizontal',
+  vertical = 'vertical',
+}
+
+export enum DisplayValues {
+  specificationName = 'SPECIFICATION_NAME',
+  specificationValue = 'SPECIFICATION_VALUE',
+}
