--- react/typings/storefront.d.ts
@@ -1,13 +1,0 @@
-import { FunctionComponent } from 'react'
-
-declare global {
-  interface StorefrontFunctionComponent<P = {}> extends FunctionComponent<P> {
-    getSchema?(props: P): object
-    schema?: object
-  }
-
-  interface StorefrontComponent<P = {}, S = {}> extends Component<P, S> {
-    getSchema?(props: P): object
-    schema: object
-  }
-}
